fix: Rename `components/next` to `components/layout`
