+++ frontend/src/app/components/client-provider.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { BaseLayout } from '@/app/components/next/base-layout'
+import { BaseLayout } from '@/app/components/layout/base-layout'
 import { env } from '@/config/environment'
 import { getDeployments } from '@/deployments/deployments'
 import { UseInkathonProvider } from '@scio-labs/use-inkathon'
