+++ frontend/src/app/components/layout/center-body.tsx
