+++ frontend/src/app/components/layout/hot-toast-config.tsx
